+++ src/assets/scripts/spacing.ts
@@ -1,0 +1,23 @@
+type spacingTypes = {
+  [key: string]: string;
+};
+
+const spacing: spacingTypes = {
+  'size-spacing-6xs': 'var(--size-50)',
+  'size-spacing-5xs': 'var(--size-100)',
+  'size-spacing-4xs': 'var(--size-150)',
+  'size-spacing-3xs': 'var(--size-200)',
+  'size-spacing-2xs': 'var(--size-250)',
+  'size-spacing-xs': 'var(--size-300)',
+  'size-spacing-sm': 'var(--size-400)',
+  'size-spacing-md': 'var(--size-500)',
+  'size-spacing-lg': 'var(--size-600)',
+  'size-spacing-xl': 'var(--size-700)',
+  'size-spacing-2xl': 'var(--size-800)',
+  'size-spacing-3xl': 'var(--size-900)',
+  'size-spacing-4xl': 'var(--size-1000)',
+  'size-spacing-5xl': 'var(--size-1100)',
+  'size-spacing-6xl': 'var(--size-1200)',
+};
+
+export default spacing;