+++ src/components/sidenav/use-sidenav.ts
@@ -1,19 +1,26 @@
+import type { SetupContext } from 'vue';
+
+import type { SidenavPropTypes, SidenavEmitTypes } from './sidenav';
+
 interface MouseEventWithCtrl extends MouseEvent {
   ctrlKey: boolean;
 }
 
 interface Redirect {
-  isThirdPartyLink: boolean;
+  openInNewTab: boolean;
+  isAbsoluteURL: boolean;
   link: string;
 }
 
-export const useSidenav = () => {
+export const useSidenav = (props: SidenavPropTypes, emit: SetupContext<SidenavEmitTypes>['emit']) => {
   const handleRedirect = (e: MouseEventWithCtrl, redirect: Redirect) => {
-    if (redirect && e.ctrlKey && e.button === 0) {
-      if (redirect.isThirdPartyLink) {
+    if (redirect) {
+      if (redirect.openInNewTab) {
         window.open(redirect.link, '_blank');
-      } else {
+      } else if (redirect.isAbsoluteURL) {
         location.href = redirect.link;
+      } else {
+        emit('route-push', redirect.link);
       }
     }
   };